remove uncessary file test=develop
